Remove LIST_MSG implementation on cloud files
Remove LIST_MSG implementation on cloud files and remove the
cloud_device representation. The use of this message is to list all
devices registered on the cloud. The use of LIST_MSG is not necessary
on VirtualThing because this repository is a implementation of KNoT
architecture 1.0 or/and KNoT architecture 2.0 (see
https://knot.cesar.org.br).