Refactor structure to store the device settings
Remove the struct conf_files from the conf_parameters header and add a
struct on the device header to store the settings required by the
device_start function. Also, add a new field to store the credentials
globally on the device source. These changes are necessary since the
device_start function parameters must be accessible by every part of
code which includes the device header.