Fix cloud exchange and queue names
The KNoT Fog solution was rebuilt to be more lightweight, resilient and
performative. Considering it's now reintegrated to the platform, this
patch updates the cloud capabilities to send/receive message to it
instead of connector directly. Also, the queue names were updated to
improve traceability by adding a "thingd" prefix to them.