Update KNoT Cloud SDK and KNoT Protocol version
The main change on these libraries is the restructuring of the config
and schema structures. The name of config was changed to event and, now,
config is the representation of schema and event.
So, this commit updates the version of KNoT Cloud SDK and KNoT Protocol
and adapt the code to be compilable and functional.